feat: Puts node.js and brower.js back in and works with index.ts files to export programmatic access
