+++ .eslintrc.json
@@ -1,3 +1,3 @@
 {
-  "extends": "./build/node.js"
+  "extends": "./node.js"
 }
