+++ .github/workflows/qa.yml
@@ -21,6 +21,10 @@ jobs:
     - name: Install dependencies
       run: |
         npm install
+    # Build has to happen before for linting to work
+    - name: Build
+      run: |
+        npx roboter build
     - name: Run roboter
       run: |
         npx roboter
