+++ lib/betterRules/index.ts
@@ -1,0 +1,12 @@
+import { BetterRulesRecord } from './BetterRulesRecord';
+import { compile } from './compile';
+import { usePlugin } from './usePlugin';
+
+export {
+  compile,
+  usePlugin
+};
+
+export type {
+  BetterRulesRecord
+};
