+++ lib/index.ts
@@ -1,0 +1,7 @@
+import * as betterRules from './betterRules';
+import * as defaultConfig from './defaultConfig';
+
+export {
+  betterRules,
+  defaultConfig
+};
