+++ node.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const defaultEslintConfig = require('./build/lib/defaultConfig');
+
+module.exports = defaultEslintConfig;
